Add `Environment.from_folder()` to mimic envdir
Now our software is able to use the same format that `envdir` uses to
load folders where each file is a variable and the content of the file
is the value.

It's also possible to delete stuff from the environment too. If you do
something like this:

    >>> env = Environment.from_folder('/etc/envdir/yipit')
    >>> del env['KEY']

It will actually delete the file called `KEY` inside of the environment
path pointed in the constructor.